Refactor StopWatch.tsx into hooks and utils to follow Single Responsiblity
